+++ src/main/java/generation/rencapp/api/DepartamentoRestController.java
@@ -1,55 +1,6 @@
-/*package generation.rencapp.API;
-
-import generation.rencapp.models.Departamento;
-import generation.rencapp.service.DepartamentoService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("api/departamentos")
-@CrossOrigin("*")
-
-public class DepartamentoRestController {
-
-    @Autowired
-    private DepartamentoService departamentoService;
-
-    @GetMapping("/lista")
-    public List <Departamento> buscarLista() {
-        return departamentoService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Departamento> buscarPorId(@PathVariable Long id) {
-        return new ResponseEntity<>(departamentoService.findById(id), HttpStatus.OK);
-    }
-
-    @PostMapping("/crear")
-    public Departamento crearDepartamento(@RequestBody Departamento departamento) {
-        return departamentoService.save(departamento);
-    }
-
-    @DeleteMapping("/eliminar/{id}")
-    public String eliminarDepartamento(@PathVariable Long id) {
-        departamentoService.deleteById(id);
-        return "Departamento eliminado";
-    }
-
-    @PutMapping("/editar/{id}")
-    public Departamento editarDepartamento(@PathVariable Long id, @RequestBody Departamento departamento) {
-        departamento.setId(id);
-        return departamentoService.save(departamento);
-    }
-}*/
-
 package generation.rencapp.api;
-
 import generation.rencapp.models.Departamento;
-import generation.rencapp.service.DepartamentoService;
+import generation.rencapp.services.DepartamentoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
