+++ src/main/java/generation/rencapp/api/ServicioRestController.java
@@ -1,48 +1,8 @@
-/*package generation.rencapp.API;
-import generation.rencapp.models.Servicio;
-import generation.rencapp.service.ServicioService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
-@RestController
-@RequestMapping("api/servicios")
-@CrossOrigin("*")
-public class ServicioRestController {
-    @Autowired
-    private ServicioService servicioService;
-
-    @GetMapping("/lista")
-    public ResponseEntity<List<Servicio>> buscarLista() {
-        return new ResponseEntity<>(servicioService.findAll(), HttpStatus.OK);
-    }
-
-    @GetMapping("/{id}")
-    public Servicio buscarPorId(@PathVariable Long id) {
-        return servicioService.findById(id);}
-    @PostMapping("/crear")
-    public Servicio crearServicio(@RequestBody Servicio Servicio) {
-        return servicioService.save(Servicio);
-    }
-    @DeleteMapping("/eliminar/{id}")
-    public String eliminarServicio(@PathVariable Long id) {
-        servicioService.deleteById(id);
-        return "Servicio eliminado";
-    }
-    @PutMapping("/editar/{id}")
-    public Servicio editarServicio(@PathVariable Long id, @RequestBody Servicio Servicio) {
-        Servicio.setId(id);
-        return servicioService.save(Servicio);
-    }
-}*/
-
 package generation.rencapp.api;
-
 import generation.rencapp.models.Departamento;
 import generation.rencapp.models.Servicio;
-import generation.rencapp.service.DepartamentoServiceImpl;
-import generation.rencapp.service.ServicioServiceImpl;
+import generation.rencapp.services.DepartamentoServiceImpl;
+import generation.rencapp.services.ServicioServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
