+++ src/main/java/generation/rencapp/repositories/DepartamentoRepository.java
@@ -1,0 +1,9 @@
+package generation.rencapp.repositories;
+
+import generation.rencapp.models.Departamento;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
+}
