--- src/main/java/generation/rencapp/repository/DepartamentoRepository.java
@@ -1,9 +1,0 @@
-package generation.rencapp.repository;
-
-import generation.rencapp.models.Departamento;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
-}
