--- src/main/java/generation/rencapp/repository/ServicioRepository.java
@@ -1,14 +1,0 @@
-package generation.rencapp.repository;
-
-import generation.rencapp.models.Servicio;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ServicioRepository extends JpaRepository<Servicio, Long> {
-    List<Servicio> findByDepartamentoId(Long departamentoId);
-
-    boolean existsByDepartamentoId(Long departamentoId);
-}
